refactor : Spring Rest Docs 설정 중복 제거
