--- src/test/java/louie/hanse/shareplate/integration/ChatRoomMemberIntegrationTest.java
@@ -1,84 +1,0 @@
-package louie.hanse.shareplate.integration;
-
-import static io.restassured.RestAssured.given;
-import static org.springframework.http.HttpHeaders.AUTHORIZATION;
-import static org.springframework.http.HttpHeaders.CONNECTION;
-import static org.springframework.http.HttpHeaders.CONTENT_LENGTH;
-import static org.springframework.http.HttpHeaders.DATE;
-import static org.springframework.http.HttpHeaders.HOST;
-import static org.springframework.http.HttpHeaders.TRANSFER_ENCODING;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.removeHeaders;
-import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
-import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
-
-import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
-import io.restassured.http.ContentType;
-import io.restassured.specification.RequestSpecification;
-import java.util.Collections;
-import louie.hanse.shareplate.jwt.JwtProvider;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
-
-@DisplayName("채팅방의 회원 기능 통합 테스트")
-@DisplayNameGeneration(ReplaceUnderscores.class)
-@ExtendWith({RestDocumentationExtension.class})
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-class ChatRoomMemberIntegrationTest {
-
-    @LocalServerPort
-    int port;
-    
-    RequestSpecification documentationSpec;
-
-    @Autowired
-    JwtProvider jwtProvider;
-
-    @BeforeEach
-    void setUp(RestDocumentationContextProvider restDocumentation) {
-        RestAssured.port = port;
-        documentationSpec = new RequestSpecBuilder()
-            .addFilter(
-                documentationConfiguration(restDocumentation)
-                    .operationPreprocessors()
-                    .withRequestDefaults(
-                        prettyPrint(),
-                        removeHeaders(HOST, CONTENT_LENGTH))
-                    .withResponseDefaults(
-                        prettyPrint(),
-                        removeHeaders(CONTENT_LENGTH, CONNECTION, DATE, TRANSFER_ENCODING, "Keep-Alive",
-                            HttpHeaders.VARY))
-            )
-            .build();
-    }
-
-    @Test
-    void 참가한_채팅방을_나간다() {
-        String accessToken = jwtProvider.createAccessToken(2370842997L);
-
-        given(documentationSpec)
-            .contentType(ContentType.JSON)
-            .filter(document("chatRoomMember-chatRoom-exit-delete"))
-            .header(AUTHORIZATION, accessToken)
-            .body(Collections.singletonMap("chatRoomId", 1))
-
-            .when()
-            .delete("/chatroom-members")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-}
