+++ src/test/java/louie/hanse/shareplate/integration/InitIntegrationTest.java
@@ -1,7 +1,5 @@
 package louie.hanse.shareplate.integration;
 
-import static io.restassured.RestAssured.given;
-import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.HttpHeaders.CONNECTION;
 import static org.springframework.http.HttpHeaders.CONTENT_LENGTH;
 import static org.springframework.http.HttpHeaders.DATE;
@@ -9,41 +7,30 @@ import static org.springframework.http.HttpHeaders.HOST;
 import static org.springframework.http.HttpHeaders.TRANSFER_ENCODING;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.removeHeaders;
-import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
 
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
-import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
-import louie.hanse.shareplate.jwt.JwtProvider;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.restdocs.RestDocumentationContextProvider;
 import org.springframework.restdocs.RestDocumentationExtension;
 
-@DisplayName("참가 기능 통합 테스트")
 @DisplayNameGeneration(ReplaceUnderscores.class)
 @ExtendWith({RestDocumentationExtension.class})
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class EntryIntegrationTest {
+public class InitIntegrationTest {
 
     @LocalServerPort
     int port;
 
-    RequestSpecification documentationSpec;
-
-    @Autowired
-    JwtProvider jwtProvider;
+    protected RequestSpecification documentationSpec;
 
     @BeforeEach
     void setup(RestDocumentationContextProvider restDocumentation) {
@@ -57,44 +44,9 @@ public class EntryIntegrationTest {
                         removeHeaders(HOST, CONTENT_LENGTH))
                     .withResponseDefaults(
                         prettyPrint(),
-                        removeHeaders(CONTENT_LENGTH, CONNECTION, DATE, TRANSFER_ENCODING,
-                            "Keep-Alive",
+                        removeHeaders(CONTENT_LENGTH, CONNECTION, DATE, TRANSFER_ENCODING, "Keep-Alive",
                             HttpHeaders.VARY))
             )
             .build();
-    }
-
-    @Test
-    void 특정_회원이_특정_쉐어에_참가_한다() {
-        String accessToken = jwtProvider.createAccessToken(2355841047L);
-
-        given(documentationSpec)
-            .filter(document("entry-share"))
-            .contentType(ContentType.JSON)
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 2)
-
-            .when()
-            .post("/shares/{id}/entry")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-
-    @Test
-    void 특정_회원의_특정_쉐어_참가를_취소한다() {
-        String accessToken = jwtProvider.createAccessToken(2355841047L);
-
-        given(documentationSpec)
-            .filter(document("entry-cancel"))
-            .contentType(ContentType.JSON)
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 1)
-
-            .when()
-            .delete("/shares/{id}/entry")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
     }
 }
