--- src/test/java/louie/hanse/shareplate/integration/KeywordIntegrationTest.java
@@ -1,105 +1,0 @@
-package louie.hanse.shareplate.integration;
-
-import static io.restassured.RestAssured.given;
-import static org.springframework.http.HttpHeaders.AUTHORIZATION;
-import static org.springframework.http.HttpHeaders.CONNECTION;
-import static org.springframework.http.HttpHeaders.CONTENT_LENGTH;
-import static org.springframework.http.HttpHeaders.DATE;
-import static org.springframework.http.HttpHeaders.HOST;
-import static org.springframework.http.HttpHeaders.TRANSFER_ENCODING;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.removeHeaders;
-import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
-
-import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
-import io.restassured.http.ContentType;
-import io.restassured.specification.RequestSpecification;
-import java.util.Map;
-import louie.hanse.shareplate.jwt.JwtProvider;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
-
-@DisplayName("키워드 통합 테스트")
-@DisplayNameGeneration(ReplaceUnderscores.class)
-@ExtendWith(RestDocumentationExtension.class)
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-class KeywordIntegrationTest {
-
-    @LocalServerPort
-    int port;
-
-    RequestSpecification documentationSpec;
-
-    @Autowired
-    JwtProvider jwtProvider;
-
-    @BeforeEach
-    void setUp(RestDocumentationContextProvider restDocumentation) {
-        RestAssured.port = port;
-        documentationSpec = new RequestSpecBuilder()
-            .addFilter(
-                documentationConfiguration(restDocumentation)
-                    .operationPreprocessors()
-                    .withRequestDefaults(
-                        prettyPrint(),
-                        removeHeaders(HOST, CONTENT_LENGTH))
-                    .withResponseDefaults(
-                        prettyPrint(),
-                        removeHeaders(CONTENT_LENGTH, CONNECTION, DATE, TRANSFER_ENCODING,
-                            "Keep-Alive",
-                            HttpHeaders.VARY))
-            )
-            .build();
-    }
-
-    @Test
-    void 새로_생긴_쉐어에_대한_알림을_받을_수_있도록_키워드를_등록한다() {
-        String accessToken = jwtProvider.createAccessToken(2370842997L);
-
-        Map<String, Object> requestBody = Map.of(
-            "location", "강남역",
-            "latitude", 37.498095,
-            "longitude", 127.027610,
-            "keyword", "떡볶이"
-        );
-
-        given(documentationSpec)
-            .contentType(ContentType.JSON)
-            .header(AUTHORIZATION, accessToken)
-            .body(requestBody)
-
-            .when()
-            .post("/keywords")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-
-    @Test
-    void 등록한_키워드를_삭제한다() {
-        String accessToken = jwtProvider.createAccessToken(2370842997L);
-
-        given()
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 1)
-
-            .when()
-            .delete("/keywords/{id}")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-}
