--- src/test/java/louie/hanse/shareplate/integration/WishIntegrationTest.java
@@ -1,107 +1,0 @@
-package louie.hanse.shareplate.integration;
-
-import static io.restassured.RestAssured.given;
-import static org.springframework.http.HttpHeaders.AUTHORIZATION;
-import static org.springframework.http.HttpHeaders.CONNECTION;
-import static org.springframework.http.HttpHeaders.CONTENT_LENGTH;
-import static org.springframework.http.HttpHeaders.DATE;
-import static org.springframework.http.HttpHeaders.HOST;
-import static org.springframework.http.HttpHeaders.TRANSFER_ENCODING;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.removeHeaders;
-import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
-import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
-
-import io.restassured.RestAssured;
-import io.restassured.builder.RequestSpecBuilder;
-import io.restassured.http.ContentType;
-import io.restassured.specification.RequestSpecification;
-import louie.hanse.shareplate.jwt.JwtProvider;
-import net.minidev.json.JSONObject;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
-
-@DisplayName("위시 기능 통합 테스트")
-@DisplayNameGeneration(ReplaceUnderscores.class)
-@ExtendWith({RestDocumentationExtension.class})
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class WishIntegrationTest {
-
-    @LocalServerPort
-    int port;
-
-    RequestSpecification documentationSpec;
-
-    @Autowired
-    JwtProvider jwtProvider;
-
-    @BeforeEach
-    void setup(RestDocumentationContextProvider restDocumentation) {
-        RestAssured.port = port;
-        documentationSpec = new RequestSpecBuilder()
-            .addFilter(
-                documentationConfiguration(restDocumentation)
-                    .operationPreprocessors()
-                    .withRequestDefaults(
-                        prettyPrint(),
-                        removeHeaders(HOST, CONTENT_LENGTH))
-                    .withResponseDefaults(
-                        prettyPrint(),
-                        removeHeaders(CONTENT_LENGTH, CONNECTION, DATE, TRANSFER_ENCODING,
-                            "Keep-Alive",
-                            HttpHeaders.VARY))
-            )
-            .build();
-    }
-
-    @Test
-    void 특정_회원이_특정_쉐어를_위시로_등록한다() {
-        String accessToken = jwtProvider.createAccessToken(2355841047L);
-
-        JSONObject requestParam = new JSONObject();
-        requestParam.put("shareId", 1);
-
-        given(documentationSpec)
-            .filter(document("wish-list-share"))
-            .contentType(ContentType.JSON)
-            .header(AUTHORIZATION,accessToken)
-            .body(requestParam)
-
-            .when()
-            .post("/wish-list")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-
-    @Test
-    void 특정_회원이_특정_쉐어의_위시를_취소한다() {
-        String accessToken = jwtProvider.createAccessToken(2355841047L);
-
-        JSONObject requestParam = new JSONObject();
-        requestParam.put("shareId", 2);
-
-        given(documentationSpec)
-            .filter(document("wish-list-cancel-share"))
-            .contentType(ContentType.JSON)
-            .header(AUTHORIZATION,accessToken)
-            .body(requestParam)
-
-            .when()
-            .delete("/wish-list")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-}
