+++ src/test/java/louie/hanse/shareplate/integration/login/LoginIntegrationTest.java
@@ -1,4 +1,4 @@
-package louie.hanse.shareplate.integration;
+package louie.hanse.shareplate.integration.login;
 
 import static io.restassured.RestAssured.given;
 import static org.springframework.http.HttpHeaders.CONNECTION;
@@ -15,53 +15,22 @@ import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
+import louie.hanse.shareplate.integration.InitIntegrationTest;
 import louie.hanse.shareplate.jwt.JwtProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.restdocs.RestDocumentationContextProvider;
-import org.springframework.restdocs.RestDocumentationExtension;
 
 @DisplayName("로그인 기능 통합 테스트")
-@DisplayNameGeneration(ReplaceUnderscores.class)
-@ExtendWith({RestDocumentationExtension.class})
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class LoginIntegrationTest {
-
-    @LocalServerPort
-    int port;
-
-    RequestSpecification documentationSpec;
+class LoginIntegrationTest extends InitIntegrationTest {
 
     @Autowired
     JwtProvider jwtProvider;
-
-    @BeforeEach
-    void setup(RestDocumentationContextProvider restDocumentation) {
-        RestAssured.port = port;
-        documentationSpec = new RequestSpecBuilder()
-            .addFilter(
-                documentationConfiguration(restDocumentation)
-                    .operationPreprocessors()
-                    .withRequestDefaults(
-                        prettyPrint(),
-                        removeHeaders(HOST, CONTENT_LENGTH))
-                    .withResponseDefaults(
-                        prettyPrint(),
-                        removeHeaders(CONTENT_LENGTH, CONNECTION, DATE, TRANSFER_ENCODING,
-                            "Keep-Alive",
-                            HttpHeaders.VARY))
-            )
-            .build();
-    }
 
     @Test
     void 로그인_회원을_로그아웃_한다() {
